requirements.txt
cassandra-driver==3.14.0
kafka-python==1.4.3
six==1.11.0
